LayoutTests/accessibility/ios-simulator/description-list-expected.txt

term 1
def 1
This tests that description lists are exposed on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS term.isInDescriptionListTerm is true
PASS def.isInDescriptionListDetail is true
PASS successfullyParsed is true

TEST COMPLETE
